+++ packages/web-components/src/_docs/design-system/color-explorer/component-types.ts
@@ -1,0 +1,6 @@
+export enum ComponentTypes {
+  backplate = 'backplate',
+  text = 'text',
+  form = 'form',
+  sample = 'sample',
+}
