+++ packages/web-components/src/_docs/design-system/color-explorer/components/color-block.ts
@@ -30,7 +30,7 @@ import {
   neutralStrokeRest,
   SwatchRGB,
 } from '../../../../index-rollup';
-import { ComponentTypes } from '../app';
+import { ComponentTypes } from '../component-types';
 import { AppSwatch } from './swatch';
 
 AppSwatch;
